refactor(transport): split 742-line system.ts god file
Resolves the #25 god-file item. AGENTS.md §3 mandates splitting service
files past ~300 lines; system.ts was 742 with 19 exports across 4
unrelated concerns.

- dashboard.ts   (228) — static asset serving + in-memory cache
- sdk-proxy.ts    (92) — tools/project/agents/providers/MCP/LSP proxy
- filesystem.ts  (196) — file browser (list/read/glob/readAbs)
- system.ts      (262) — core system handlers (status/connect-info/
  health/token-rotate) + shared utils (extractDirectory, getIP) + a
  barrel re-exporting the moved symbols

Pure move — handler bodies byte-identical, public export surface
unchanged (same 19 names/types), so routes.ts / sessions.ts /
projects.ts need ZERO import changes. Dropped one dead import
(validateToken) and hoisted two mid-file imports (ESM-equivalent).
src/transport/AGENTS.md updated to document the new layout.

Zero behavior change. tsc clean, 558/0. Independent fresh-context
review: APPROVE, no blocking issues (module-level dashboard cache
semantics verified preserved).

Stacked on #30 → #26 (modifies handlers/ alongside the stack).
Refs #25.
