+++ src/transport/http/handlers/sdk-proxy.ts
@@ -1,0 +1,92 @@
+import type { RouteContext } from "../routes"
+import { json, jsonError } from "../middlewares/json"
+import { CORS_HEADERS } from "../middlewares/cors"
+import { extractDirectory } from "./system"
+import type {
+  Agent,
+  LspStatus,
+  McpStatus,
+} from "@opencode-ai/sdk"
+
+// ─── SDK proxy handlers ──────────────────────────────────────────────────────
+
+export async function listTools({ url, deps }: RouteContext): Promise<Response> {
+  const dirParam = extractDirectory(url)
+  if (dirParam === null)
+    return jsonError("INVALID_DIRECTORY", "Internal error: the dashboard sent an invalid directory path. Try refreshing the page; if it persists, report at https://github.com/lesquel/open-remote-control/issues.", 400, CORS_HEADERS)
+  const result = await deps.client.tool.ids({ query: { ...dirParam } })
+  return json(result.data ?? [], 200, CORS_HEADERS)
+}
+
+export async function getProject({ deps }: RouteContext): Promise<Response> {
+  return json(
+    {
+      project: deps.project,
+      directory: deps.directory,
+      worktree: deps.worktree,
+    },
+    200,
+    CORS_HEADERS,
+  )
+}
+
+export async function listAgents({ url, deps }: RouteContext): Promise<Response> {
+  const dirParam = extractDirectory(url)
+  if (dirParam === null)
+    return jsonError("INVALID_DIRECTORY", "Internal error: the dashboard sent an invalid directory path. Try refreshing the page; if it persists, report at https://github.com/lesquel/open-remote-control/issues.", 400, CORS_HEADERS)
+  try {
+    const result = await deps.client.app.agents({ query: { ...dirParam } })
+    const agents: Array<Agent> = result.data ?? []
+    return json({ agents }, 200, CORS_HEADERS)
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err)
+    deps.logger.error("SDK call failed: app.agents", { error: message })
+    return jsonError("SDK_ERROR", "SDK call failed", 500, CORS_HEADERS)
+  }
+}
+
+export async function listProviders({ url, deps }: RouteContext): Promise<Response> {
+  const dirParam = extractDirectory(url)
+  if (dirParam === null)
+    return jsonError("INVALID_DIRECTORY", "Internal error: the dashboard sent an invalid directory path. Try refreshing the page; if it persists, report at https://github.com/lesquel/open-remote-control/issues.", 400, CORS_HEADERS)
+  try {
+    const result = await deps.client.provider.list({ query: { ...dirParam } })
+    return json(result.data ?? {}, 200, CORS_HEADERS)
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err)
+    deps.logger.error("SDK call failed: provider.list", { error: message })
+    return jsonError("SDK_ERROR", "SDK call failed", 500, CORS_HEADERS)
+  }
+}
+
+export async function getMcpStatus({ url, deps }: RouteContext): Promise<Response> {
+  const dirParam = extractDirectory(url)
+  if (dirParam === null)
+    return jsonError("INVALID_DIRECTORY", "Internal error: the dashboard sent an invalid directory path. Try refreshing the page; if it persists, report at https://github.com/lesquel/open-remote-control/issues.", 400, CORS_HEADERS)
+  try {
+    const result = await deps.client.mcp.status({ query: { ...dirParam } })
+    const servers: Record<string, McpStatus> = result.data ?? {}
+    return json({ servers }, 200, CORS_HEADERS)
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err)
+    deps.logger.error("SDK call failed: mcp.status", { error: message })
+    return jsonError("SDK_ERROR", "SDK call failed", 500, CORS_HEADERS)
+  }
+}
+
+// ─── LSP status ──────────────────────────────────────────────────────────────
+
+export async function getLspStatus({ url, deps }: RouteContext): Promise<Response> {
+  const dirParam = extractDirectory(url)
+  if (dirParam === null)
+    return jsonError("INVALID_DIRECTORY", "Internal error: the dashboard sent an invalid directory path. Try refreshing the page; if it persists, report at https://github.com/lesquel/open-remote-control/issues.", 400, CORS_HEADERS)
+  try {
+    const result = await deps.client.lsp.status({ query: { ...dirParam } })
+    const clients: Array<LspStatus> = result.data ?? []
+    return json({ clients }, 200, CORS_HEADERS)
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err)
+    deps.logger.error("SDK call failed: lsp.status", { error: message })
+    return jsonError("SDK_ERROR", "SDK call failed", 500, CORS_HEADERS)
+  }
+}
